requirements.txt
fastapi
uvicorn
python-dotenv
groq
openai
google-generativeai
requests
httpx
beautifulsoup4
psycopg
psycopg-binary
python-multipart
pydantic
APScheduler
duckduckgo-search
google-api-python-client
google-auth
google-auth-httplib2
pymongo
motor
python-jose
passlib
bcrypt
pyparsing
dnspython
email-validator
sqlalchemy
sqlalchemy
psycopg2-binary
ddgs